xyz/requirements.txt
flask
fastapi
uvicorn
torch
torchvision
transformers
flask-cors
scikit-learn
pandas
numpy
seaborn
matplotlib
minisom
PySyft
tensorflow
stable-baselines3
google-fitbit-api-client
streamlit